carrito y detalle
puse el boton a detalle y ahora anda el agregar al carrito

+++ borrar.php
@@ -1,14 +1,19 @@
 <?php
-    include 'conexion.php';
+include 'conexion.php';
+session_start();
+
     if($_SERVER["REQUEST_METHOD"] === "POST" ) {
-        /*$idCarrito = $_POST['id_carrito'];
-        $sql = "DELETE FROM carrito WHERE id = '$idCarrito'";
-        if ($conexion->query($sql) === TRUE) {
-            echo "OK";
-        } else {
-            echo "Error al " . $conexion->error;
-        }
-        $conexion->close();*/
+        //borrar del carrito
+            if(isset($_POST['idCarrito'])){
+                $idCarrito = $_POST['idCarrito'];
+                $sql = "DELETE FROM carrito WHERE id = '$idCarrito'";
+                if ($conexion->query($sql) === TRUE) {
+                    echo "productp";
+                } else {
+                    echo "Error: " . $conexion->error;
+                }
+            }
+        
         $idProducto = $_POST['eliminarProducto'];
         $sql_deleteProducto = "DELETE FROM producto WHERE id = '$idProducto'";
         if ($conexion->query($sql_deleteProducto) === TRUE) {
